Hardware SPI ChipSelect pin can be any digital output
If you specify a Chip Select pin other than A2, A2 will not be set as an output by the SPI class in Spark Core firmware.  Then you can use it as an analog input if you wish.  Also reordered the SPI mode configuration to before SPI.begin() so as not to glitch the bus immediately after enabling SPI.